More compact data type for storing values

+++ src/data_store.cpp
@@ -1,5 +1,5 @@
 #include "data_store.h"
 
-int32_t pm_sensor::DataStore::float_to_stored(float value) {
-  return (int32_t)(value * 10.0);
+int16_t pm_sensor::DataStore::float_to_stored(float value) {
+  return (int16_t)(value * 10.0);
 }
